@@ -14,7 +14,7 @@ group 'io.github.newohoo.project'
 version '2025.1.2'
 
 java {
-    sourceCompatibility = JavaVersion.VERSION_1_8
+    sourceCompatibility = JavaVersion.VERSION_17
 }
 
 repositories {
